s_library 2.5 Сервіс адаптовано до WL1.0. Оновлено інсталяційну модель, керування контентом переведоно до системних бібліотек.

+++ app/services/library/models/install_model.php
@@ -15,9 +15,9 @@ class install
 	public $admin_ico = 'fa-book';
 	public $version = "2.5";
 
-	public $options = array('useGroups' => 1, 'articleMultiGroup' => 0, 'useAvailability' => 0, 'folder' => 'library', 'articleOrder' => 'position DESC', 'groupOrder' => 'position ASC');
-	public $options_type = array('useGroups' => 'bool', 'articleMultiGroup' => 'bool', 'useAvailability' => 'bool', 'folder' => 'text', 'articleOrder' => 'text', 'groupOrder' => 'text');
-	public $options_title = array('useGroups' => 'Наявність груп', 'articleMultiGroup' => 'Мультигрупи (1 стаття більше ніж 1 група)', 'useAvailability' => 'Використання доступності', 'folder' => 'Папка для зображень/аудіо', 'articleOrder' => 'Сортування товарів', 'groupOrder' => 'Сортування груп');
+	public $options = array('useGroups' => 1, 'articleMultiGroup' => 0, 'folder' => 'library', 'articleOrder' => 'position DESC', 'groupOrder' => 'position ASC');
+	public $options_type = array('useGroups' => 'bool', 'articleMultiGroup' => 'bool', 'folder' => 'text', 'articleOrder' => 'text', 'groupOrder' => 'text');
+	public $options_title = array('useGroups' => 'Наявність груп', 'articleMultiGroup' => 'Мультигрупи (1 стаття більше ніж 1 група)', 'folder' => 'Папка для зображень/аудіо', 'articleOrder' => 'Сортування товарів', 'groupOrder' => 'Сортування груп');
 	public $options_admin = array (
 					'word:articles_to_all' => 'статтей',
 					'word:article_to' => 'До статті',
@@ -42,35 +42,34 @@ class install
 			$query = "CREATE TABLE IF NOT EXISTS `{$this->table_service}_groups` (
 						  `id` int(11) NOT NULL AUTO_INCREMENT,
 						  `wl_alias` int(11) NOT NULL,
-						  `alias` text NOT NULL,
-						  `parent` int(11),
-						  `position` int(11) NOT NULL,
-						  `photo` text NOT NULL,
-						  `active` tinyint(1) NOT NULL,
+						  `alias` text,
+						  `parent` int(11) DEFAULT NULL,
+						  `position` int(11) DEFAULT NULL,
+						  `active` tinyint(1) DEFAULT NULL,
 						  `author_add` int(11) NOT NULL,
 						  `date_add` int(11) NOT NULL,
 						  `author_edit` int(11) NOT NULL,
 						  `date_edit` int(11) NOT NULL,
 						  PRIMARY KEY (`id`),
-						  UNIQUE KEY `id` (`id`)
-						) ENGINE=InnoDB  DEFAULT CHARSET=utf8;";
+						  UNIQUE KEY `id` (`id`),
+						  KEY `wl_alias` (`wl_alias`),
+						  KEY `parent` (`parent`),
+						  KEY `position` (`position`),
+						  KEY `active` (`active`)
+						) ENGINE=InnoDB DEFAULT CHARSET=utf8;";
 			$this->db->executeQuery($query);
 
 			if($this->options['articleMultiGroup'] > 0)
 			{
 				$query = "CREATE TABLE IF NOT EXISTS `{$this->table_service}_article_group` (
 						  `article` int(11) NOT NULL,
-						  `group` int(11) NOT NULL
+						  `group` int(11) NOT NULL,
+						  KEY `article` (`wl_alias`),
+						  KEY `group` (`parent`)
 						) ENGINE=InnoDB  DEFAULT CHARSET=utf8;";
 				$this->db->executeQuery($query);
 			}
 		}
-
-		$path = IMG_PATH.$this->options['folder'].'/groups';
-		$path = substr($path, strlen(SITE_URL));
-		if(!is_dir($path))
-			mkdir($path, 0777);
-
 		return true;
 	}
 
@@ -80,21 +79,21 @@ class install
 		if(!empty($articles))
 		{
 			$this->db->deleteRow($this->table_service.'_articles', $alias, 'wl_alias');
-			$this->db->deleteRow($this->table_service.'_groups', $alias, 'wl_alias');
-
-			foreach ($articles as $article) {
-				$this->db->deleteRow($this->table_service.'_article_group', $article->id, 'article');
-			}
+			if($this->options['useGroups'] > 0)
+				$this->db->deleteRow($this->table_service.'_groups', $alias, 'wl_alias');
+			
+			if($this->options['articleMultiGroup'] > 0)
+				foreach ($articles as $article) {
+					$this->db->deleteRow($this->table_service.'_article_group', $article->id, 'article');
+				}
 		}
 
-		$groups = $this->db->getAllDataByFieldInArray($this->table_service.'_groups', $alias, 'wl_alias');
-		if(!empty($groups))
-			$this->db->deleteRow($this->table_service.'_groups', $alias, 'wl_alias');
-
-		$path = IMG_PATH.$this->options['folder'].'/groups';
-		$path = substr($path, strlen(SITE_URL));
-		if(is_dir($path))
-			$this->removeDirectory($path);
+		if($this->options['useGroups'] > 0)
+		{
+			$groups = $this->db->getAllDataByFieldInArray($this->table_service.'_groups', $alias, 'wl_alias');
+			if(!empty($groups))
+				$this->db->deleteRow($this->table_service.'_groups', $alias, 'wl_alias');
+		}
 
 		return true;
 	}
@@ -108,25 +107,30 @@ class install
 			$query = "CREATE TABLE IF NOT EXISTS `{$this->table_service}_groups` (
 						  `id` int(11) NOT NULL AUTO_INCREMENT,
 						  `wl_alias` int(11) NOT NULL,
-						  `alias` text NOT NULL,
-						  `parent` int(11),
-						  `position` int(11) NOT NULL,
-						  `photo` int(11) NOT NULL,
-						  `active` tinyint(1) NOT NULL,
+						  `alias` text,
+						  `parent` int(11) DEFAULT NULL,
+						  `position` int(11) DEFAULT NULL,
+						  `active` tinyint(1) DEFAULT NULL,
 						  `author_add` int(11) NOT NULL,
 						  `date_add` int(11) NOT NULL,
 						  `author_edit` int(11) NOT NULL,
 						  `date_edit` int(11) NOT NULL,
 						  PRIMARY KEY (`id`),
-						  UNIQUE KEY `id` (`id`)
-						) ENGINE=InnoDB  DEFAULT CHARSET=utf8;";
+						  UNIQUE KEY `id` (`id`),
+						  KEY `wl_alias` (`wl_alias`),
+						  KEY `parent` (`parent`),
+						  KEY `position` (`position`),
+						  KEY `active` (`active`)
+						) ENGINE=InnoDB DEFAULT CHARSET=utf8;";
 			$this->db->executeQuery($query);
 		}
 		if($option == 'articleMultiGroup' AND $value > 0)
 		{
 			$query = "CREATE TABLE IF NOT EXISTS `{$this->table_service}_article_group` (
 						  `article` int(11) NOT NULL,
-						  `group` int(11) NOT NULL
+						  `group` int(11) NOT NULL,
+						  KEY `article` (`wl_alias`),
+						  KEY `group` (`parent`)
 						) ENGINE=InnoDB  DEFAULT CHARSET=utf8;";
 			$this->db->executeQuery($query);
 
@@ -154,18 +158,21 @@ class install
 		$query = "CREATE TABLE IF NOT EXISTS `{$this->table_service}_articles` (
 					  `id` int(11) NOT NULL AUTO_INCREMENT,
 					  `wl_alias` int(11) NOT NULL,
-					  `alias` text NOT NULL,
-					  `group` int(11) NOT NULL,
-					  `availability` int(11) NOT NULL,
-					  `active` tinyint(1) NOT NULL,
-					  `position` int(11) NOT NULL,
+					  `alias` text,
+					  `group` int(11) DEFAULT NULL,
+					  `active` tinyint(1) DEFAULT NULL,
+					  `position` int(11) DEFAULT NULL,
 					  `author_add` int(11) NOT NULL,
 					  `date_add` int(11) NOT NULL,
 					  `author_edit` int(11) NOT NULL,
 					  `date_edit` int(11) NOT NULL,
 					  PRIMARY KEY (`id`),
-					  UNIQUE KEY `id` (`id`)
-					) ENGINE=InnoDB  DEFAULT CHARSET=utf8;";
+					  UNIQUE KEY `id` (`id`),
+					  KEY `wl_alias` (`wl_alias`),
+					  KEY `group` (`group`),
+					  KEY `active` (`active`),
+					  KEY `position` (`position`)
+					) ENGINE=InnoDB DEFAULT CHARSET=utf8";
 		$this->db->executeQuery($query);
 
 		return true;
@@ -180,15 +187,6 @@ class install
 			$this->db->executeQuery("DROP TABLE IF EXISTS {$this->table_service}_groups");
 		}
 		return true;
-	}
-
-	private function removeDirectory($dir) {
-	    if ($objs = glob($dir."/*")) {
-	       foreach($objs as $obj) {
-	         is_dir($obj) ? $this->removeDirectory($obj) : unlink($obj);
-	       }
-	    }
-	    rmdir($dir);
 	}
 	
 }
